+++ R/rc_helpers.R
@@ -44,7 +44,7 @@ rc_update_weather <- function(dt = NULL){
     dt <- copy(dt)
     
     # Check inputs
-    checkmate::assert_data_table(dt, nrows = 12)
+    checkmate::assert_data_table(dt, min.rows = 12)
     req <- c("month", "W_TEMP_MEAN_MONTH", "W_PREC_SUM_MONTH")
     checkmate::assert_names(colnames(dt), must.include = req)
     checkmate::assert(